+++ lib/WorseReflection/Tests/Unit/Core/Inference/FrameTest.php
@@ -4,7 +4,7 @@ namespace Phpactor\WorseReflection\Tests\Unit\Core\Inference;
 
 use Phpactor\WorseReflection\Core\Inference\Assignments;
 use PHPUnit\Framework\TestCase;
-use Phpactor\WorseReflection\Core\Inference\ConcreteFrame;
+use Phpactor\WorseReflection\Core\Inference\Frame\ConcreteFrame;
 use Phpactor\WorseReflection\Core\Inference\Frame;
 use Phpactor\WorseReflection\Core\Inference\LocalAssignments;
 use Phpactor\WorseReflection\Core\Inference\PropertyAssignments;
